set up a more reliable way to execute external process

+++ ExperimentRunner/src/main/scala/peal/runner/actor/Z3CallerActor.scala
@@ -18,12 +18,21 @@ class Z3CallerActor(memoryLimit: Long) extends Actor {
     //TODO this is not advisable, google for solution on dealing with huge messages
     case input: String =>
       val tmp = File.createTempFile("z3file", "")
+      val execTmp = File.createTempFile("callz3", "")
+      execTmp.setExecutable(true)
       FileUtil.writeToFile(tmp.getAbsolutePath, input)
-//      example to set path
-//      Process(Seq("bash", "-c", "ulimit", "-v", memoryLimit.toString)) #&& Process(Seq("z3", "-nw", "-smt2", tmp.getAbsolutePath), None, "PATH" -> z3Path) ! processLogger
-    //TODO generate a shell script the execute it in a process
-      Process(Seq("bash", "-c", "ulimit", "-v", memoryLimit.toString)) #&& Process(Seq("z3", "-nw", "-smt2", tmp.getAbsolutePath)) ! processLogger
+      val script = "#!/bin/sh\nulimit -v " + memoryLimit + "\nz3 -nw -smt2 " + tmp.getAbsolutePath + "\n"
+      FileUtil.writeToFile(execTmp.getAbsolutePath, script)
+      //      Process(Seq("bash", "-c", "ulimit", "-v", memoryLimit.toString)) #&& Process(Seq("z3", "-nw", "-smt2", tmp.getAbsolutePath)) ! processLogger
+      val returnCode = Seq(execTmp.getAbsolutePath) ! processLogger
+      //      println(execTmp.getAbsolutePath + " " + returnCode)
+
+      if (returnCode != 0 && returnCode != 1) {
+        throw new RuntimeException("Kill due to memory restriction")
+      }
+
       tmp.delete()
+      execTmp.delete()
 
       sender ! resultList.mkString("\n")
   }
